feat: Add support for blueprint reading
Added basic support for reading data from blueprints.  Detected some
problems with the stat preprocessor that are not easy to solve.
Typically the preprocessor will remove any shapes that are touching the
edge of the boundary box.  However detected that a number and a shape
are touching the boundary box.  Meaning we have no hope of removing that
shape with our current criteria.