feat: Improve algorithm for stat value detection
Improved the alogrithm for detecting stat values. We no longer remove
edges detected at the edge of the bounding box from the image. Instead
the neural network is now trained to detect this shapes and report them.
We then ignore these images when computing the value of the image.
Also refactored main.py to use a delegate. New commands will be added in
the command delegate instead.
Also changed the ordering ot STAT_OPTIONS to reflect the order they
would appear in an image.